+++ lib/chores/models/status.dart
