+++ lib/pages/home_page.dart
@@ -1,6 +1,6 @@
-import 'package:chorebuddies_flutter/tasks/task_view.dart';
+import 'package:chorebuddies_flutter/chores/chore_view.dart';
 import 'package:flutter/material.dart';
-import 'package:chorebuddies_flutter/tasks/mock_task_list.dart';
+import 'package:chorebuddies_flutter/chores/mock_chore_list.dart';
 
 class HomePage extends StatelessWidget {
   const HomePage({Key? key}) : super(key: key);
@@ -8,7 +8,7 @@ class HomePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     // Example data for the ListView
-    final items = mockTasks;
+    final items = mockChores;
 
     return Padding(
       padding: const EdgeInsets.all(8.0),
@@ -16,7 +16,7 @@ class HomePage extends StatelessWidget {
         child: ListView.builder(
           itemCount: items.length,
           itemBuilder: (context, index) {
-            return TaskView(task: items[index]);
+            return ChoreView(chore: items[index]);
           },
         ),
       ),
